+++ Sunshine-Equipment/build.gradle
@@ -6,6 +6,7 @@ plugins {
 
 group 'com.cognizant'
 version '0.0.1-SNAPSHOT'
+//version '1.0.0-SNAPSHOT'
 
 sourceCompatibility = 1.8
