+++ Sunshine-Equipment/src/main/java/com/cognizant/equipmentcrud/EquipmentCrudApplication.java
@@ -1,0 +1,16 @@
+package com.cognizant.equipmentcrud;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+@EnableResourceServer
+public class EquipmentCrudApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EquipmentCrudApplication.class, args);
+    }
+}
