+++ Sunshine-Equipment/src/main/java/com/cognizant/equipmentcrud/config/SecurityConfig.java
@@ -1,0 +1,10 @@
+package com.cognizant.equipmentcrud.config;
+
+import org.springframework.boot.autoconfigure.security.oauth2.client.EnableOAuth2Sso;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@EnableOAuth2Sso
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+}
