+++ Sunshine-Equipment/src/main/java/com/cognizant/equipmentcrud/repository/EquipmentRepository.java
@@ -1,0 +1,9 @@
+package com.cognizant.equipmentcrud.repository;
+
+import com.cognizant.equipmentcrud.model.Equipment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
+}
