--- Sunshine-Equipment/src/main/java/controller/ControllerExceptionHandler.java
@@ -1,23 +1,0 @@
-package controller;
-
-import org.springframework.hateoas.VndErrors;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.context.request.WebRequest;
-
-@RestControllerAdvice
-@RequestMapping(produces = "application/vnd.error+json")
-public class ControllerExceptionHandler {
-
-    @ExceptionHandler(value = {IllegalArgumentException.class})
-    @ResponseStatus(HttpStatus.UNPROCESSABLE_ENTITY)
-    public ResponseEntity<VndErrors> illegalArgumentException(IllegalArgumentException e, WebRequest request) {
-        VndErrors error = new VndErrors(request.toString(), e.getMessage());
-        ResponseEntity<VndErrors> responseEntity = new ResponseEntity<>(error, HttpStatus.UNPROCESSABLE_ENTITY);
-        return responseEntity;
-    }
-}
