--- Sunshine-Equipment/src/main/java/repository/EquipmentRepository.java
@@ -1,9 +1,0 @@
-package repository;
-
-import model.Equipment;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
-}
