+++ Sunshine-Equipment/src/main/resources/data.sql
@@ -1,17 +1,17 @@
 INSERT INTO Equipment (equipment_type, office_id, assigned, user_id) values
 ('Windows Laptop', 1, 'TRUE', 1);
 
-INSERT INTO Equipment (equipment_type, office_id, assigned, comment) values
+INSERT INTO Equipment (equipment_type, office_id, assigned, comments) values
 ('Windows Laptop', 1, 'FALSE', 'not being used');
 
 INSERT INTO Equipment (equipment_type, office_id, assigned, user_id) values
 ('Windows Laptop', 1, 'TRUE', 2);
 
 INSERT INTO Equipment (equipment_type, office_id, assigned, user_id) values
-('Windows Laptop', 2, 'TRUE', 1);
+('Windows Laptop', 2, 'TRUE', 3);
 
-INSERT INTO Equipment (equipment_type, office_id, assigned, comment) values
+INSERT INTO Equipment (equipment_type, office_id, assigned, comments) values
 ('Windows Laptop', 2, 'FALSE', 'not being used');
 
 INSERT INTO Equipment (equipment_type, office_id, assigned, user_id) values
-('Windows Laptop', 2, 'TRUE', 2);
+('Windows Laptop', 2, 'TRUE', 4);
